backend/requirements.txt
fastapi
uvicorn
pydantic
manga-ocr
Pillow